\name{VoltRonStore-package}
\alias{VoltRonStore-package}
\alias{VoltRonStore}
\docType{package}
\title{
\packageTitle{VoltRonStore}
}
\description{
\packageDescription{VoltRonStore}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{VoltRonStore}

Maintainer: \packageMaintainer{VoltRonStore}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = VoltRonStore)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{VoltRonStore}}
%\section{Documentation Index}{\packageIndices{VoltRonStore}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
